Rename post_processor to wrapper
