+++ lib/json_parser.rb
@@ -8,7 +8,7 @@ module JsonParser
   require 'concerns/options_parser'
   require 'json_parser/version'
   require 'json_parser/crawler'
-  require 'json_parser/post_processor'
+  require 'json_parser/wrapper'
   require 'json_parser/fetcher'
   require 'json_parser/class_methods'
 end
