+++ lib/json_parser/wrapper.rb
