+++ spec/support/shared_examples/wrapper.rb
